improve document around custom markdown plugin and add escape sequence utils
