+++ src/app/plugins/markdown/block/index.ts
@@ -1,0 +1,1 @@
+export * from './parser';
